-> Issue left (Debug Issue on modify Update Medicine In RV for First Time) (May be DiffUtils) -> Only Privacy Policy and Invoice Left (Wholesaler)

+++ app/src/main/java/com/talhachaudhry/jpharmaappfyp/admin/ManageMedicineActivity.java
@@ -2,14 +2,18 @@ package com.talhachaudhry.jpharmaappfyp.admin;
 
 import androidx.annotation.IdRes;
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import androidx.lifecycle.ViewModelProvider;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -23,6 +27,7 @@ import com.talhachaudhry.jpharmaappfyp.databinding.ActivityManageMedicineBinding
 import com.talhachaudhry.jpharmaappfyp.login_details.Login;
 import com.talhachaudhry.jpharmaappfyp.models.ManageMedicineModel;
 import com.talhachaudhry.jpharmaappfyp.view_models.ManageMedicineViewModel;
+import com.talhachaudhry.jpharmaappfyp.wholesaler.fragments.AnimationFragment;
 
 import java.util.Objects;
 
@@ -39,9 +44,13 @@ public class ManageMedicineActivity extends AppCompatActivity implements OnViewM
         setContentView(binding.getRoot());
         Objects.requireNonNull(getSupportActionBar()).hide();
         viewModel = new ViewModelProvider(this).get(ManageMedicineViewModel.class);
+        if (viewModel.getIsLoading()) {
+            openAnimation();
+        }
         adapter = new ManageMedicineRecyclerAdapter(this, this);
         viewModel.getManageMedicineViewModelMutableLiveData().observe(this, manageMedicineModels ->
                 adapter.submitList(manageMedicineModels));
+        viewModel.getAllMedicinesSet().observe(this, aBoolean -> runOnUiThread(() -> getSupportFragmentManager().popBackStack()));
         binding.manageMedicineRv.setAdapter(adapter);
         binding.backBtn.setOnClickListener(view -> onBackPressed());
         binding.addMedicine.setOnClickListener(view ->
@@ -65,7 +74,27 @@ public class ManageMedicineActivity extends AppCompatActivity implements OnViewM
 
     @Override
     public void deleteMedicine(ManageMedicineModel model) {
-        viewModel.deleteMedicine(model);
+        confirmationDialog(model);
+    }
+
+    void confirmationDialog(ManageMedicineModel model) {
+        new AlertDialog.Builder(this)
+                .setTitle("Delete Medicine")
+                .setIcon(R.drawable.delete_ic)
+                .setMessage("Are you sure you want to delete " + model.getName())
+                .setCancelable(true)
+                .setPositiveButton("Yes", (dialogInterface, i) -> {
+                    viewModel.deleteMedicine(model);
+                    dialogInterface.dismiss();
+                })
+                .setNegativeButton("NO", (dialogInterface, i) -> dialogInterface.dismiss())
+                .show();
+    }
+
+    void openAnimation() {
+        new Handler(Looper.getMainLooper()).post(() ->
+                openFragment(AnimationFragment.newInstance(R.raw.loader_animation, ""),
+                R.id.animation_container));
     }
 
     @Override
